--- CodingChallenge/CodingChallenge/Articles.swift
@@ -1,14 +1,0 @@
-//
-//  Articles.swift
-//  CodingChallenge
-//
-//  Created by margaret on 2019-06-25.
-//  Copyright © 2019 margaret. All rights reserved.
-//
-
-import Foundation
-
-struct Articles: Codable {
-    //var children: [Article]
-    var data: Article
-}
