+++ CodingChallenge/CodingChallenge/Config.swift
@@ -1,0 +1,11 @@
+//
+//  Config.swift
+//  CodingChallenge
+//
+//  Created by margaret on 2019-06-26.
+//  Copyright © 2019 margaret. All rights reserved.
+//
+
+import Foundation
+
+let DATA_SOURCE = "https://www.reddit.com/r/swift.json"
