+++ CodingChallenge/CodingChallenge/Models/Article.swift
@@ -1,0 +1,9 @@
+//
+//  File.swift
+//  CodingChallenge
+//
+//  Created by margaret on 2019-06-26.
+//  Copyright © 2019 margaret. All rights reserved.
+//
+
+import Foundation
